Temp rename capitalized files
